backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
reportlab==4.2.5
google-cloud-storage==2.18.2
supabase==2.9.1
resend==2.4.0
sendgrid==6.11.0
python-multipart==0.0.12
